scripts/manifestProcessing_uart/requirements.txt
awscli
cryptoauthlib >= 20190831
cryptography >= 2.7
pyasn1_modules == 0.1.5
pytz >= 2018.9
python_jose
boto3
requests_aws4auth
hexrec
pyserial
